sfzh.py -> stars.py + split histories for clarity
